requirements.txt
streamlit==1.35.0
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.4.0
python-dotenv==1.0.0
Pillow==10.1.0
PyPDF2==3.0.1
pandas==2.1.3
python-docx==0.8.11
openpyxl==3.11.0
